Fix compile error on x64 config

+++ shell/disassembler/code_data_provider.h
@@ -38,7 +38,7 @@ namespace dpr {
 
     class DynamicCodeDataProvider : public CodeDataProvider {
     public:
-        DynamicCodeDataProvider(uint32_t base, HANDLE process);
+        DynamicCodeDataProvider(intptr_t base, HANDLE process);
 
         uint8_t get8(uint32_t off) const override;
         uint16_t get16(uint32_t off) const override;
@@ -46,7 +46,7 @@ namespace dpr {
         uint64_t get64(uint64_t off) const override;
 
     private:
-        uint32_t base_;
+        intptr_t base_;
         HANDLE process_;
     };
